+++ packages/e2e/cypress/e2e/app/tableCalculation.cy.ts
@@ -131,7 +131,7 @@ describe('Table calculations', () => {
         cy.get(".mantine-Select-input[value='is']").click();
         cy.contains('greater than').click(); // If the type is string, this option will not be available and it will fail when running the query
 
-        cy.findByPlaceholderText('Enter value').clear().type('250');
+        cy.findByPlaceholderText('Enter value(s)').clear().type('250');
         cy.get('button').contains('Run query').click();
 
         cy.contains('Loading results'); // wait for results to load