+++ packages/frontend/src/components/Explorer/ResultsCard/ExplorerResults.tsx
@@ -35,17 +35,26 @@ export const ExplorerResults = memo(() => {
     const explorerColumnOrder = useExplorerContext(
         (context) => context.state.unsavedChartVersion.tableConfig.columnOrder,
     );
-    const resultsData = useExplorerContext(
-        (context) => context.queryResults.data,
+    const fetchedRows = useExplorerContext(
+        (context) => context.queryResults.fetchedRows,
+    );
+    const totalRows = useExplorerContext(
+        (context) => context.queryResults.totalRows,
+    );
+    const isFetchingRows = useExplorerContext(
+        (context) => context.queryResults.isFetchingRows,
+    );
+    const fetchMoreRows = useExplorerContext(
+        (context) => context.queryResults.fetchMoreRows,
     );
     const status = useExplorerContext((context) => {
         // Don't return context.queryResults.status because we changed from mutation to query so 'loading' as a different meaning
-        if (context.queryResults.isFetching) {
+        if (context.query.isFetching) {
             return 'loading';
-        } else if (context.queryResults.status === 'loading') {
+        } else if (context.query.status === 'loading') {
             return 'idle';
         } else {
-            return context.queryResults.status;
+            return context.query.status;
         }
     });
     const setColumnOrder = useExplorerContext(
@@ -154,7 +163,10 @@ export const ExplorerResults = memo(() => {
             <Box px="xs" py="lg">
                 <Table
                     status={status}
-                    data={resultsData?.rows || []}
+                    data={fetchedRows || []}
+                    totalRowsCount={totalRows || 0}
+                    isFetchingRows={isFetchingRows}
+                    fetchMoreRows={fetchMoreRows}
                     columns={columns}
                     columnOrder={explorerColumnOrder}
                     onColumnOrderChange={setColumnOrder}