+++ packages/frontend/src/components/common/Table/TableProvider.tsx
@@ -1,7 +1,6 @@
 import {
     getCoreRowModel,
     getExpandedRowModel,
-    getPaginationRowModel,
     useReactTable,
     type ColumnOrderState,
     type GroupingState,
@@ -19,7 +18,11 @@ import { type ProviderProps, type TableColumn } from './types';
 const rowColumn: TableColumn = {
     id: ROW_NUMBER_COLUMN_ID,
     header: '#',
-    cell: (props) => props.row.index + 1,
+    cell: (props) => {
+        const { pageIndex, pageSize } = props.table.getState().pagination;
+        const pageStartIndex = pageIndex * pageSize;
+        return pageStartIndex + props.row.index + 1;
+    },
     footer: 'Total',
     meta: {
         width: 30,
@@ -46,7 +49,14 @@ export const TableProvider: FC<React.PropsWithChildren<ProviderProps>> = ({
     children,
     ...rest
 }) => {
-    const { data, columns, columnOrder, pagination } = rest;
+    const {
+        data,
+        totalRowsCount,
+        columns,
+        columnOrder,
+        pagination,
+        fetchMoreRows,
+    } = rest;
     const [grouping, setGrouping] = useState<GroupingState>([]);
     const [columnVisibility, setColumnVisibility] = useState({});
 
@@ -117,8 +127,32 @@ export const TableProvider: FC<React.PropsWithChildren<ProviderProps>> = ({
             : [stickyRowColumn, ...stickyColumns, ...otherColumns];
     }, [hideRowNumbers, stickyColumns, otherColumns, stickyRowColumn]);
 
+    const [paginationState, setPagination] = useState({
+        pageIndex: 0,
+        pageSize: DEFAULT_PAGE_SIZE,
+    });
+    useEffect(() => {
+        // Fetch rows for next pages
+        const pageThreshold = 2;
+        const { pageIndex, pageSize } = paginationState;
+        const currentPageRowCount = pageIndex * pageSize;
+        const nextPagesRowCount =
+            currentPageRowCount + pageSize * pageThreshold;
+        if (data.length < nextPagesRowCount) {
+            fetchMoreRows();
+        }
+    }, [data.length, fetchMoreRows, paginationState]);
+
+    const pageRows = useMemo(() => {
+        // calculate page rows from data and pagination state
+        const { pageIndex, pageSize } = paginationState;
+        const start = pageIndex * pageSize;
+        const end = start + pageSize;
+        return data.slice(start, end);
+    }, [data, paginationState]);
+
     const table = useReactTable({
-        data,
+        data: pageRows,
         columns: visibleColumns,
         state: {
             grouping,
@@ -130,12 +164,16 @@ export const TableProvider: FC<React.PropsWithChildren<ProviderProps>> = ({
                     ...stickyColumns.map((c) => c.id || ''),
                 ],
             },
+            pagination: paginationState,
         },
         enablePinning: true,
         onColumnVisibilityChange: setColumnVisibility,
         onColumnOrderChange: setTempColumnOrder,
         getCoreRowModel: getCoreRowModel(),
-        getPaginationRowModel: getPaginationRowModel(),
+        manualPagination: true,
+        rowCount: totalRowsCount,
+        pageCount: Math.ceil(totalRowsCount / paginationState.pageSize),
+        onPaginationChange: setPagination,
         onGroupingChange: setGrouping,
         groupedColumnMode: false,
         getExpandedRowModel: getExpandedRowModel(),